ajustado a parte de jogar.
