+++ core/src/br/edu/utfpr/SacoMoedaProcessor.java
@@ -1,19 +1,32 @@
 package br.edu.utfpr;
 
+import com.badlogic.gdx.Input;
 import com.badlogic.gdx.InputProcessor;
 
-public class BulletsProcessor implements InputProcessor {
-    BulletsProcessor(){
+public class SacoMoedaProcessor implements InputProcessor {
+    public boolean isSPressed = false, isWPressed=false;
+    SacoMoedaProcessor() {
         ShowDoMilhao.addInputProcessor(this);
     }
     @Override
     public boolean keyDown(int keycode) {
-        BulletController.ref.addNewBullet(MainScreen.ref.moeda.getX()+150,MainScreen.ref.moeda.getY()+40);
+        if(keycode == Input.Keys.S){
+            isSPressed=true;
+        }
+        if(keycode == Input.Keys.W){
+            isWPressed = true;
+        }
         return false;
     }
 
     @Override
     public boolean keyUp(int keycode) {
+        if(keycode == Input.Keys.S){
+            isSPressed=false;
+        }
+        if(keycode == Input.Keys.W){
+            isWPressed = false;
+        }
         return false;
     }
